Move networking conf into a dedicated pytest plugin
